Move pkg/brokerapi to contrib/pkg/brokerapi
