+++ contrib/pkg/brokerapi/openservicebroker/open_service_broker_client.go
@@ -28,10 +28,10 @@ import (
 	"time"
 
 	"github.com/golang/glog"
-	"github.com/kubernetes-incubator/service-catalog/pkg/brokerapi"
+	"github.com/kubernetes-incubator/service-catalog/contrib/pkg/brokerapi"
 	"github.com/kubernetes-incubator/service-catalog/pkg/util"
 
-	"github.com/kubernetes-incubator/service-catalog/pkg/brokerapi/openservicebroker/constants"
+	"github.com/kubernetes-incubator/service-catalog/contrib/pkg/brokerapi/openservicebroker/constants"
 )
 
 const (
